+++ npm/kriging-rs-wasm/src/spacetime/ordinary.ts
@@ -59,11 +59,13 @@ export class SpaceTimeOrdinaryKriging {
         packed.spatialSill,
         packed.spatialRange,
         packed.spatialShape,
+        packed.spatialShape2,
         packed.temporalType,
         packed.temporalNugget,
         packed.temporalSill,
         packed.temporalRange,
         packed.temporalShape,
+        packed.temporalShape2,
         packed.k1,
         packed.k2,
         packed.k3