+++ npm/kriging-rs-wasm/src/spacetime/simple.ts
@@ -57,11 +57,13 @@ export class SpaceTimeSimpleKriging {
         packed.spatialSill,
         packed.spatialRange,
         packed.spatialShape,
+        packed.spatialShape2,
         packed.temporalType,
         packed.temporalNugget,
         packed.temporalSill,
         packed.temporalRange,
         packed.temporalShape,
+        packed.temporalShape2,
         packed.k1,
         packed.k2,
         packed.k3